pyproject.toml
[project]
name = "fyp-backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "authlib>=1.6.0",
    "boto3>=1.38.37",
    "deepgram-sdk>=4.6.0",
    "fastapi>=0.115.12",
    "httpx>=0.28.1",
    "itsdangerous>=2.2.0",
    "nltk>=3.9.1",
    "openai>=1.88.0",
    "pillow>=11.2.1",
    "psycopg2-binary==2.9.10",
    "pymupdf>=1.26.1",
    "python-docx>=1.2.0",
    "python-dotenv>=1.1.0",
    "python-jose>=3.5.0",
    "python-multipart>=0.0.20",
    "python-pptx>=1.0.2",
    "qdrant-client>=1.15.0",
    "redis>=6.2.0",
    "reportlab==4.0.7",
    "uvicorn>=0.34.3",
]

[dependency-groups]
dev = [
    "psycopg2>=2.9.10",
]
